Add new packet type, move JIT to individual sim
The slowdown of moving JIT from stream interface to
simulator is big. (4.5s -> 7.3s of matrix)
However, this helps the timing information correct.
Also, phase needs timing information from simulators,
If we skips the packets, the information of cache would not
be correct. Thus, we move JIT to individual simulators.