Fix bug when parsing line number. Add section to ET_Program
The main reason of this bug is taht .plt and other sections are
also counted as a part of program and the algorithm tries to find
its line number which causes an incorrect line mapping.
The other part of the bug is solved by reset/clear the
walk_count_vector after each search. If one wants to have a
walk count for whole execution, we might need to add an
g_walk_count_vector variable for that.