Fix double-free error of stopping QEMU while running phase tracing
The real cause of this error message is not double free.
When QEMU stops by pressing ctrl+a  x, it calls exit(0) directly.
The lack of graceful shutdown makes other threads use null
resources(pointers).
We add a flag when QEMU stops and sleep for a while to solve this.