Critical fix: x86_64 SMP issue on event tracing
The critical bug was that the execution of syscall may not(usually)
be on the same core of the process/thread which issued the syscall.
This caused the tracing on mmap, execv, etc. fail.

The solution is to implement the kernel space "current" variable
(macro) in QEMU. The revision of this commit aims to do so by
introducing a new function get_syscall_user_thread().
All kernel event callback functions are re-written to use
target process (the thread which issues syscall) instead of
per-core state variable. The result is safe code on SMP systems.

Lastly, to complete the task, device driver is required to pass new
data to VPMU; the VPMU_MMAP_THREAD_SIZE.