Critical fix: phase results does not contain line mapping
This was caused by wrong address mapping. The start-end address of
executable segment was overwrited by the value of data segment.
This problem is solved by adding condition to mmap_updated_flag.
The result is that only executable segment is recorded.

The bug was from commit "c40b8c1" where it refactors this code.

Future work is to build a mapping table of data segment as well.