[Fix] Remove old style VPMU reset when executing a new program
This is originally used for event tracing.
Back to the time where VPMU does not trace process so well and use
this to reset all timers when a process starts.
Now, this will cause a problem when monitoring multiple processes.

The new mechanism is recording all counters and time in each process
instead of system wide.